final-pjt-back/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
numpy==1.26.4
outcome==1.3.0.post0
pillow==10.3.0
pycparser==2.22
PySocks==1.7.1
requests==2.31.0
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
